+++ README.md
@@ -1,25 +1,68 @@
-# Build a Trivia Game
+This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-For this project, you will use React to build a trivia game powered by [the Open Trivia Database](https://opentdb.com/).
+## Available Scripts
 
-This project does not include any wireframes: you will have to design the game yourself. One option is to present the user with a list of categories, and then quiz them within that category, keeping score, but that's not the only way you could do this. Use your creativity!
+In the project directory, you can run:
 
-## Requirements
+### `npm start`
 
-- Your React application should be broken up into 3 or more components.
-- Your application should have two or more "screens." In a traditional web application, we'd think of these as different pages or views.
-- Your application uses React's state to manage data.
-- Your application uses Open Trivia Database's API.
-- Your application is styled.
+Runs the app in the development mode.<br />
+Open [http://localhost:3000](http://localhost:3000) to view it in the browser.
 
-## About the API
+The page will reload if you make edits.<br />
+You will also see any lint errors in the console.
 
-You can learn about the Open Trivia Database API at [this API config page](https://opentdb.com/api_config.php).
+### `npm test`
 
-Some things you are likely to want to do:
+Launches the test runner in the interactive watch mode.<br />
+See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
 
-- Get list of categories: https://opentdb.com/api_category.php
-- Get 10 questions: https://opentdb.com/api.php?amount=10
-- If you add a session token you won't receive the same question twice
-  - Get a token: https://opentdb.com/api_token.php?command=request
-  - Add a token to a request: `&token=YOURTOKENHERE`
+### `npm run build`
+
+Builds the app for production to the `build` folder.<br />
+It correctly bundles React in production mode and optimizes the build for the best performance.
+
+The build is minified and the filenames include the hashes.<br />
+Your app is ready to be deployed!
+
+See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
+
+### `npm run eject`
+
+**Note: this is a one-way operation. Once you `eject`, you can’t go back!**
+
+If you aren’t satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+
+Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you’re on your own.
+
+You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
+
+## Learn More
+
+You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
+
+To learn React, check out the [React documentation](https://reactjs.org/).
+
+### Code Splitting
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/code-splitting
+
+### Analyzing the Bundle Size
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size
+
+### Making a Progressive Web App
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app
+
+### Advanced Configuration
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/advanced-configuration
+
+### Deployment
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/deployment
+
+### `npm run build` fails to minify
+
+This section has moved here: https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify
